requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
crispy-bootstrap4==2022.1
Django==4.2.4
django-bootstrap4==23.2
django-crispy-forms==2.0
djangorestframework==3.14.0
Pillow==10.0.0
pytz==2023.3
soupsieve==2.4.1
sqlparse==0.4.4
tzdata==2023.3
